+++ apps/sla/tests/conftest.py
@@ -1,6 +1,6 @@
 import pytest
 
-from apps.sla.models import SLOPolicy
+from apps.sla.models import SLAPolicy, SLOPolicy
 
 
 @pytest.fixture(autouse=True)
@@ -11,7 +11,7 @@ def enable_db_access_for_all_tests(db):
 @pytest.fixture(autouse=True)
 def clean_policies():
     """
-    Clean SLO policies before and after every test
+    Clean SLA/SLO policies before and after every test
 
         * before so it is not mixed with some leftovers
         * after so we do not leave any leftovers
@@ -19,4 +19,5 @@ def clean_policies():
     if we do it only before or only after the tests might behave differently
     when run in batch than when run alone so better to be safe then sorry
     """
+    SLAPolicy.objects.all().delete()
     SLOPolicy.objects.all().delete()