+++ collectors/ps_constants/core.py
@@ -6,7 +6,7 @@ from django.conf import settings
 from django.db import transaction
 from requests_gssapi import HTTPSPNEGOAuth
 
-from apps.sla.models import SLO, SLOPolicy
+from apps.sla.models import SLA, SLO, SLAPolicy, SLOPolicy
 from apps.trackers.models import JiraBugIssuetype
 from collectors.cveorg.models import Keyword
 from osidb.models import SpecialConsiderationPackage
@@ -45,6 +45,19 @@ def sync_special_consideration_packages(sc_packages):
     for name in sc_packages:
         package = SpecialConsiderationPackage(name=name)
         package.save()
+
+
+@transaction.atomic
+def sync_sla_policies(sla_policies):
+    """
+    Sync SLA policy data
+    """
+    SLA.objects.all().delete()
+    SLAPolicy.objects.all().delete()
+    for order, policy_desc in enumerate(sla_policies):
+        # In SLA policies order is important so it is passed down to the model
+        policy = SLAPolicy.create_from_description(policy_desc, order)
+        policy.save()
 
 
 @transaction.atomic