+++ osidb/management/commands/sync_ps_constants.py
@@ -14,6 +14,7 @@ class Command(BaseCommand):
         (
             cveorg_keywords,
             sc_packages,
+            sla_policies,
             slo_policies,
             jira_bug_issuetype,
         ) = collect_step_1_fetch()
@@ -21,6 +22,7 @@ class Command(BaseCommand):
         collect_step_2_sync(
             cveorg_keywords,
             sc_packages,
+            sla_policies,
             slo_policies,
             jira_bug_issuetype,
         )